+++ frontend/app/tests/unit/setup-files/setup.ts
@@ -85,6 +85,27 @@ beforeAll(() => {
     createI18n: () => ({}),
   }));
 
+  vi.mock('vue-router', () => {
+    const route = ref({
+      query: {},
+    });
+
+    return {
+      useRoute: vi.fn().mockReturnValue(route),
+      useRouter: vi.fn().mockImplementation(() => ({
+        currentRoute: route,
+        push: vi.fn(({ query }) => {
+          set(route, { ...get(route), query });
+          return true;
+        }),
+      })),
+      createRouter: vi.fn().mockImplementation(() => ({
+        beforeEach: vi.fn(),
+      })),
+      createWebHashHistory: vi.fn(),
+    };
+  });
+
   vi.mock('@/store/websocket', () => ({
     useWebsocketStore: () => ({
       connected: ref(false),
